fix(eslint-plugin): support eslint@5
