+++ packages/eslint-plugin/typings/eslint-rules.d.ts
@@ -830,17 +830,3 @@ declare module 'eslint/lib/rules/prefer-const' {
   >;
   export = rule;
 }
-
-declare module 'eslint/lib/rules/utils/ast-utils' {
-  import { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
-
-  const utils: {
-    getNameLocationInGlobalDirectiveComment(
-      sourceCode: TSESLint.SourceCode,
-      comment: TSESTree.Comment,
-      name: string,
-    ): TSESTree.SourceLocation;
-  };
-
-  export = utils;
-}
